+++ lib/core/alice_chopper_response_interceptor.dart
@@ -10,11 +10,16 @@ import 'alice_core.dart';
 
 class AliceChopperInterceptor extends chopper.ResponseInterceptor
     with chopper.RequestInterceptor {
+  /// AliceCore instance
   final AliceCore aliceCore;
 
-  AliceChopperInterceptor(this.aliceCore);
+  /// Creates instance of chopper interceptor
+  AliceChopperInterceptor(this.aliceCore)
+      : assert(aliceCore != null, "aliceCore can't be null");
 
+  /// Creates hashcode based on request
   int getRequestHashCode(BaseRequest baseRequest) {
+    assert(baseRequest != null, "baseReqeust can't be null");
     int hashCodeSum = 0;
     if (baseRequest.url != null) {
       hashCodeSum += baseRequest.url.hashCode;
@@ -35,8 +40,10 @@ class AliceChopperInterceptor extends chopper.ResponseInterceptor
     return hashCodeSum.hashCode;
   }
 
+  /// Handles chopper request and creates alice http call
   @override
   FutureOr<chopper.Request> onRequest(chopper.Request request) async {
+    assert(request != null, "request can't be null");
     var baseRequest = await request.toBaseRequest();
     AliceHttpCall call = AliceHttpCall(getRequestHashCode(baseRequest));
     String endpoint = "";
@@ -44,7 +51,7 @@ class AliceChopperInterceptor extends chopper.ResponseInterceptor
     if (request.baseUrl == null || request.baseUrl.isEmpty) {
       List<String> split = request.url.split("/");
       if (split.length > 2) {
-        server = split[0] + "//" + split[1] + split[2];
+        server = split[1] + split[2];
       }
       if (split.length > 4) {
         endpoint = "/";
@@ -57,6 +64,7 @@ class AliceChopperInterceptor extends chopper.ResponseInterceptor
       endpoint = request.url;
       server = request.baseUrl;
     }
+
     call.method = request.method;
     call.endpoint = endpoint;
     call.server = server;
@@ -91,7 +99,9 @@ class AliceChopperInterceptor extends chopper.ResponseInterceptor
     return request;
   }
 
+  /// Handles chopper response and adds data to existing alice http call
   FutureOr<chopper.Response> onResponse(chopper.Response response) {
+    assert(response != null, "response can't be null");
     var httpResponse = AliceHttpResponse();
     httpResponse.status = response.statusCode;
     if (response.body == null) {